# openapi.yaml
...
openapi: 3.1.0
info:
  title: Olimp API
  version: 1.0.0
  description: "API мини-приложения «Олимп» в MAX: каталог олимпиад, льготы вузов, личный план и напоминания. Защищённые методы требуют заголовок Authorization: Bearer <accessToken> из POST /auth/max или POST /auth/test."
servers:
  - url: https://{domain}/api
    description: Публичный сервер
    variables:
      domain:
        default: olimp.example.ru
        description: Домен из APP_DOMAIN
  - url: http://localhost:8080/api
    description: "Локальный запуск: docker compose up"
...
paths:
...
  /auth/max:
    post:
      operationId: postAuthMax
      tags:
        - Вход
      summary: Вход через MAX
      description: Принимает window.WebApp.initData, проверяет подпись токеном бота и выдаёт JWT на час.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                initData:
                  type: string
                  minLength: 1
                  maxLength: 16384
              required:
                - initData
              additionalProperties: false
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken:
                    type: string
                  expiresIn:
                    type: number
                    enum:
                      - 3600
                  user:
                    type: object
                    properties:
                      name:
                        type: string
                        minLength: 1
                        maxLength: 80
                      grade:
                        anyOf:
                          - type: integer
                            minimum: 1
                            maximum: 11
                          - type: "null"
                      region:
                        type: string
                        maxLength: 100
                      subjects:
                        maxItems: 35
                        type: array
                        items:
                          type: integer
                          exclusiveMinimum: 0
                          maximum: 9007199254740991
                      online:
                        type: boolean
                      onsite:
                        type: boolean
                      directions:
                        maxItems: 25
                        type: array
                        items:
                          type: string
                          pattern: ^\d{2}\.0[35]\.\d{2}$
                      universities:
                        maxItems: 20
                        type: array
                        items:
                          type: string
                          maxLength: 80
                          pattern: ^[a-z0-9]+(?:-[a-z0-9]+)*$
                      avatar:
                        default: null
                        anyOf:
                          - type: string
                            maxLength: 1400000
                          - type: "null"
                      id:
                        type: string
                        format: uuid
                        pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
                      maxUserId:
                        type: string
                      registeredAt:
                        type:
                          - string
                          - "null"
                      createdAt:
                        type: string
                    required:
                      - name
                      - grade
                      - region
                      - subjects
                      - online
                      - onsite
                      - avatar
                      - id
                      - maxUserId
                      - registeredAt
                      - createdAt
                    additionalProperties: false
                required:
                  - accessToken
                  - expiresIn
                  - user
                additionalProperties: false
        "401":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  message:
                    type: string
                required:
                  - error
                  - message
                additionalProperties: false
        "503":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  message:
                    type: string
                required:
                  - error
                  - message
                additionalProperties: false
...
  /auth/test:
    post:
      operationId: postAuthTest
      tags:
        - Вход
      summary: Вход тестовой учётной записью
      description: Доступен, только если на сервере задан TEST_ACCOUNTS. Выдаёт JWT на час.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                login:
                  type: string
                  minLength: 1
                  maxLength: 64
                password:
                  type: string
                  minLength: 1
                  maxLength: 200
              required:
                - login
                - password
              additionalProperties: false
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken:
                    type: string
                  expiresIn:
                    type: number
                    enum:
                      - 3600
                  user:
                    type: object
                    properties:
                      name:
                        type: string
                        minLength: 1
                        maxLength: 80
                      grade:
                        anyOf:
                          - type: integer
                            minimum: 1
                            maximum: 11
                          - type: "null"
                      region:
                        type: string
                        maxLength: 100
                      subjects:
                        maxItems: 35
                        type: array
                        items:
                          type: integer
                          exclusiveMinimum: 0
                          maximum: 9007199254740991
                      online:
                        type: boolean
                      onsite:
                        type: boolean
                      directions:
                        maxItems: 25
                        type: array
                        items:
                          type: string
                          pattern: ^\d{2}\.0[35]\.\d{2}$
                      universities:
                        maxItems: 20
                        type: array
                        items:
                          type: string
                          maxLength: 80
                          pattern: ^[a-z0-9]+(?:-[a-z0-9]+)*$
                      avatar:
                        default: null
                        anyOf:
                          - type: string
                            maxLength: 1400000
                          - type: "null"
                      id:
                        type: string
                        format: uuid
                        pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
                      maxUserId:
                        type: string
                      registeredAt:
                        type:
                          - string
                          - "null"
                      createdAt:
                        type: string
                    required:
                      - name
                      - grade
                      - region
                      - subjects
                      - online
                      - onsite
                      - avatar
                      - id
                      - maxUserId
                      - registeredAt
                      - createdAt
                    additionalProperties: false
                required:
                  - accessToken
                  - expiresIn
                  - user
                additionalProperties: false
        "401":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  message:
                    type: string
                required:
                  - error
                  - message
                additionalProperties: false
...
  /directions:
    get:
      operationId: getDirections
      tags:
        - Поступление
      summary: Направления подготовки
      security: []
      parameters:
        - schema:
            type: string
            maxLength: 100
          in: query
          name: q
          required: false
        - schema:
            type: string
            enum:
              - "true"
              - "false"
          in: query
          name: popular
          required: false
        - schema:
            minItems: 1
            maxItems: 50
            type: array
            items:
              type: string
              maxLength: 80
              pattern: ^[a-z0-9]+(?:-[a-z0-9]+)*$
          in: query
          name: universities
          required: false
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      type: object
                      properties:
                        code:
                          type: string
                          pattern: ^\d{2}\.0[35]\.\d{2}$
                        name:
                          type: string
                        educationLevel:
                          type: string
                          enum:
                            - bachelor
                            - specialist
                        ugsnCode:
                          type: string
                        ugsnName:
                          type: string
                        popular:
                          type: boolean
                        aliases:
                          type: array
                          items:
                            type: string
                        note:
                          type:
                            - string
                            - "null"
                        egeSubjects:
                          type: array
                          items:
                            type: string
                        subjectsCore:
                          type: array
                          items:
                            type: string
                        subjectsRelated:
                          type: array
                          items:
                            type: string
                        universityCount:
                          type: integer
                          minimum: -9007199254740991
                          maximum: 9007199254740991
                        programCount:
                          type: integer
                          minimum: -9007199254740991
                          maximum: 9007199254740991
                      required:
                        - code
                        - name
                        - educationLevel
                        - ugsnCode
                        - ugsnName
                        - popular
                        - aliases
                        - note
                        - egeSubjects
                        - subjectsCore
                        - subjectsRelated
                        - universityCount
                        - programCount
                      additionalProperties: false
                required:
                  - items
                additionalProperties: false
...
  /universities:
    get:
      operationId: getUniversities
      tags:
        - Поступление
      summary: Вузы с льготами
      security: []
      parameters:
        - schema:
            minItems: 1
            maxItems: 50
            type: array
            items:
              type: string
              pattern: ^\d{2}\.0[35]\.\d{2}$
          in: query
          name: directions
          required: false
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      type: object
                      properties:
                        slug:
                          type: string
                          maxLength: 80
                          pattern: ^[a-z0-9]+(?:-[a-z0-9]+)*$
                        name:
                          type: string
                        city:
                          type: string
                        fullName:
                          type:
                            - string
                            - "null"
                        seriesCount:
                          type: integer
                          minimum: -9007199254740991
                          maximum: 9007199254740991
                        olympiadCount:
                          type: integer
                          minimum: -9007199254740991
                          maximum: 9007199254740991
                        programCount:
                          type: integer
                          minimum: -9007199254740991
                          maximum: 9007199254740991
                      required:
                        - slug
                        - name
                        - city
                        - fullName
                        - seriesCount
                        - olympiadCount
                      additionalProperties: false
                required:
                  - items
                additionalProperties: false
...
  /calendar/{file}:
    get:
      operationId: getCalendarFile
      tags:
        - План
      summary: Календарь плана в формате iCalendar
      description: Открывается по секретной ссылке из POST /me/calendar-feed без входа.
      security: []
      parameters:
        - schema:
            type: string
            pattern: ^[A-Za-z0-9_-]{43}\.ics$
          in: path
          name: file
          required: true
      responses:
        "200":
          description: Default Response
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT